# swagger.yaml
...
openapi: 3.0.0
info:
  title: DeepScraper API
  version: 1.0.0
  description: >-
    Open-source web scraping API — scrape, crawl, map, search, extract structured data, drive
    persistent browser sessions, and expose any site as a reusable endpoint.


    This file is GENERATED from the zod request schemas in `src/api/schemas`. Do not edit by hand —
    run `npm run openapi:generate`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000
    description: Local instance
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-Key
      in: header
      description: API key. Required on every /api/* route.
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - true
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
      required:
        - success
        - error
    AsyncJobAccepted:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - true
        id:
          type: string
        url:
          type: string
          description: Polling URL for this job
        status:
          type: string
      required:
        - success
        - id
        - url
        - status
...
paths:
...
  /api/extract:
    post:
      tags:
        - Extract
      summary: Async multi-URL LLM extraction
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                urls:
                  type: array
                  items:
                    type: string
                    format: uri
                  maxItems: 1000
                url:
                  type: string
                  format: uri
                prompt:
                  type: string
                  maxLength: 5000
                schema:
                  nullable: true
                limit:
                  type: integer
                  minimum: 0
                  exclusiveMinimum: true
                  maximum: 1000
                scrapeOptions:
                  type: object
                  additionalProperties:
                    nullable: true
      responses:
        '200':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobAccepted'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/extract/{id}:
    get:
      tags:
        - Extract
      summary: Poll an extract job
      security:
        - ApiKeyAuth: []
      parameters:
        - schema:
            type: string
            description: Extract job id
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/extract-auto:
    post:
      tags:
        - Extract
      summary: Self-healing extraction — derives and caches CSS selectors, re-derives on breakage
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                fields:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        minLength: 1
                        maxLength: 100
                      description:
                        type: string
                        maxLength: 500
                      type:
                        type: string
                        enum:
                          - text
                          - attribute
                          - html
                          - number
                          - list
                          - nested
                          - nested_list
                      attribute:
                        type: string
                        maxLength: 100
                      required:
                        type: boolean
                    required:
                      - name
                  minItems: 1
                  maxItems: 50
                cssSchema:
                  nullable: true
                forceReheal:
                  type: boolean
                scrapeOptions:
                  type: object
                  additionalProperties:
                    nullable: true
              required:
                - url
                - fields
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/sessions:
    post:
      tags:
        - Sessions
      summary: Create a persistent browser session
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userAgent:
                  type: string
                  maxLength: 500
                viewport:
                  type: object
                  properties:
                    width:
                      type: integer
                      minimum: 200
                      maximum: 4000
                    height:
                      type: integer
                      minimum: 200
                      maximum: 4000
                  required:
                    - width
                    - height
                initialUrl:
                  type: string
                  format: uri
                proxy:
                  type: object
                  properties:
                    server:
                      type: string
                      maxLength: 300
                    username:
                      type: string
                      maxLength: 200
                    password:
                      type: string
                      maxLength: 200
                  required:
                    - server
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Sessions
      summary: List active sessions
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sessions/{id}:
    get:
      tags:
        - Sessions
      summary: Get a session
      security:
        - ApiKeyAuth: []
      parameters:
        - schema:
            type: string
            description: Session id
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Sessions
      summary: Close a session
      security:
        - ApiKeyAuth: []
      parameters:
        - schema:
            type: string
            description: Session id
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/sessions/{id}/action:
    post:
      tags:
        - Sessions
      summary: Run an action in a session (navigate, click, scrape, …)
      security:
        - ApiKeyAuth: []
      parameters:
        - schema:
            type: string
            description: Session id
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum:
                    - navigate
                    - click
                    - type
                    - fill
                    - select
                    - scroll
                    - waitForSelector
                    - wait
                    - screenshot
                    - scrape
                    - evaluate
                    - back
                    - forward
                    - reload
                    - content
                url:
                  type: string
                  format: uri
                selector:
                  type: string
                  maxLength: 2000
                value:
                  type: string
                text:
                  type: string
                position:
                  type: number
                timeout:
                  type: integer
                  minimum: 0
                  maximum: 120000
                script:
                  type: string
                  maxLength: 20000
                fullPage:
                  type: boolean
                formats:
                  type: array
                  items:
                    type: string
                  maxItems: 6
                onlyMainContent:
                  type: boolean
                fitMarkdown:
                  type: boolean
                waitUntil:
                  type: string
                  enum:
                    - load
                    - domcontentloaded
                    - networkidle
                    - commit
              required:
                - type
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...